% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_adjacency.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Obtain the Human Ecology Grid Cell Adjacency Matrix}
\usage{
adjacency_matrix(dataset, ..., data)
}
\arguments{
\item{dataset}{the dataset descriptor (see `\code{\link[=dataset]{dataset()}})}

\item{...}{reserved for future extensions

If reference dataset \code{data} is provided, the returned matrix will match the
order of the cells in the data and drop all unused cells. This is useful when
modeling spatial autocorrelation with packages such as \code{brms}.}

\item{data}{a data frame containing a column \code{cell_id} with the cell IDs}
}
\description{
Obtain the Human Ecology Grid Cell Adjacency Matrix
}
\examples{
\dontrun{
data_descriptor <- humanEcologyGrid::dataset(resolution = 300, download = TRUE)
# cells for Pacific-Centered maps
humanEcologyGrid::grid_adjacency(data_descriptor)
}

}
