% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Human Ecology Grid dataset descriptor}
\usage{
dataset(..., resolution, download = FALSE)
}
\arguments{
\item{resolution}{grid resolution in kilometers}

\item{download}{whether to download the dataset if it is not cached

This function creates a descriptor for a specific version and grid resolution of the Human Ecology Grid Database.
The descriptor can be used to access the data and metadata.}
}
\description{
Human Ecology Grid dataset descriptor
}
\examples{
\dontrun{
data_descriptor <- humanEcologyGrid::dataset(resolution = 300, download = TRUE)
humanEcologyGrid::latest_data(data_descriptor)
}

}
