% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\name{grid_cells}
\alias{grid_cells}
\title{Obtain the Human Ecology Grid Cell Polygons}
\usage{
grid_cells(dataset, ..., split_meridian = 180, as = c("sf", "s2"))
}
\arguments{
\item{dataset}{the dataset descriptor (see `\code{\link[=dataset]{dataset()}})}

\item{...}{reserved for future extensions

The polygons can be optionally split at a given meridian for plotting or other applications
using tools that do not support spherical geometry. The default option is to split at the
antimeridian (avoids artifacts when plotting using the default projection).

The function returns a simple features (\code{sf}) object by default. To return a tibble with
\code{s2} polygons set \code{as = "s2"}.}

\item{split_meridian}{the meridian to split the cell geometry at (see details)}

\item{as}{the format to return the grid cells as (see details)}
}
\description{
Obtain the Human Ecology Grid Cell Polygons
}
\examples{
\dontrun{
data_descriptor <- humanEcologyGrid::dataset(resolution = 300, download = TRUE)
# cells for Pacific-Centered maps
humanEcologyGrid::grid_cells(data_descriptor, split_meridian = -30)
}

}
