% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_data.R
\name{latest_data}
\alias{latest_data}
\title{Obtain the Cross-Sectional (Latest) Human Ecology Grid Data}
\usage{
latest_data(dataset, ..., arrow = getOption("human.ecology.grid.arrow", FALSE))
}
\arguments{
\item{dataset}{the dataset descriptor (see `\code{\link[=dataset]{dataset()}})}

\item{...}{reserved for future extensions

Cross-sectional data contains the most recent year coverage for each data module.}

\item{arrow}{whether to load the data as an Arrow dataset or as a tibble}
}
\description{
Obtain the Cross-Sectional (Latest) Human Ecology Grid Data
}
\examples{
\dontrun{
data_descriptor <- humanEcologyGrid::dataset(resolution = 300, download = TRUE)
humanEcologyGrid::latest_data(data_descriptor)
}

}
