% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_data.R
\name{yearly_data}
\alias{yearly_data}
\title{Obtain the yearly Human Ecology Grid Data}
\usage{
yearly_data(
  dataset,
  ...,
  year,
  arrow = getOption("human.ecology.grid.arrow", FALSE)
)
}
\arguments{
\item{dataset}{the dataset descriptor (see `\code{\link[=dataset]{dataset()}})}

\item{...}{reserved for future extensions}

\item{year}{the year to load data for (numeric vector)}

\item{arrow}{whether to load the data as an Arrow dataset or as a tibble}
}
\description{
Obtain the yearly Human Ecology Grid Data
}
\examples{
\dontrun{
data_descriptor <- humanEcologyGrid::dataset(resolution = 300, download = TRUE)
humanEcologyGrid::yearly_data(data_descriptor, year = 1995)
}

}
